add some well needed documentation
