+++ src/main/java/dev/alexisok/untitledbot/Main.java
@@ -32,7 +32,7 @@ import java.util.*;
  * @author AlexIsOK
  * @since 0.0.1
  */
-public class Main {
+public final class Main {
 	
 	public static final String VERSION = "1.3.2";
 	public static final String CONFIG_PATH = Paths.get("").toAbsolutePath().toString();
