+++ src/main/java/dev/alexisok/untitledbot/ShutdownHook.java
@@ -5,7 +5,7 @@ package dev.alexisok.untitledbot;
  * @author AlexIsOK
  * @since 1.3
  */
-public class ShutdownHook extends Thread {
+public final class ShutdownHook extends Thread {
     
     @Override
     public void run() {
