--- src/main/java/dev/alexisok/untitledbot/command/CommandDoesNotExistException.java
@@ -1,8 +1,0 @@
-package dev.alexisok.untitledbot.command;
-
-/**
- * @author AlexIsOK
- * @since 0.0.1
- */
-public class CommandDoesNotExistException extends RuntimeException{
-}
