+++ src/main/java/dev/alexisok/untitledbot/command/CommandRegistrar.java
@@ -20,6 +20,24 @@ import java.util.HashMap;
 import java.util.Objects;
 
 /**
+ * This is where all commands are registered.  To register the command, you must do the following:
+ * 
+ * <ul>
+ *     <li>
+ *         Run {@link CommandRegistrar#register(String, String, Command)} to register the command.
+ *     </li>
+ *     <li>
+ *         (optional but encouraged) Add the help page: {@link Manual#setHelpPage(String, String)}
+ *     </li>
+ *     <li>
+ *         (optional) Add aliases for the command: {@link CommandRegistrar#registerAlias(String, String...)}<br>
+ *         (optional) Copy the manual pages for the aliases: {@link CommandRegistrar#registerAliasManual(String, String...)}
+ *     </li>
+ * </ul>
+ * 
+ * Commands must have a permission node.  The permission nodes must follow a regex of {@code ^[a-z]([a-z][.]?)+[a-z]$},
+ * and the command names (what the user types in on Discord) must match {@code ^[a-z0-9_-]*$}.
+ * 
  * @author AlexIsOK
  * @since 0.0.1
  */
@@ -43,7 +61,7 @@ public class CommandRegistrar {
 	 * 
 	 * @return the size of the registrar.
 	 */
-	static int registrarSize() {
+	public static int registrarSize() {
 		return REGISTRAR.size();
 	}
 	
@@ -109,6 +127,7 @@ public class CommandRegistrar {
 		if(!REGISTRAR.containsKey(commandName))
 			return null;
 		
+		//noinspection ConstantConditions
 		if(permissionNode.equals("owner") && !m.getAuthor().getId().equals(Main.OWNER_ID))
 			return null;
 			
@@ -137,7 +156,6 @@ public class CommandRegistrar {
 		try {
 			//check user permissions and guild permissions at the same time.
 			//this could be made faster.
-			@SuppressWarnings("ConstantConditions")
 			String userHas = Vault.getUserDataLocal(m.getAuthor().getId(), m.getGuild().getId(), permissionNode);
 			
 			Logger.debug("Checking the permission node of user...");
