+++ src/main/java/dev/alexisok/untitledbot/command/CoreCommands.java
@@ -1,31 +1,26 @@
 package dev.alexisok.untitledbot.command;
 
-import dev.alexisok.untitledbot.BotClass;
 import dev.alexisok.untitledbot.Main;
 import dev.alexisok.untitledbot.data.GetUserData;
 import dev.alexisok.untitledbot.logging.Logger;
 import dev.alexisok.untitledbot.modules.basic.atsomeone.AtSomeone;
 import dev.alexisok.untitledbot.modules.basic.eightball.EightBall;
+import dev.alexisok.untitledbot.modules.basic.help.Help;
+import dev.alexisok.untitledbot.modules.basic.perms.Permissions;
+import dev.alexisok.untitledbot.modules.basic.prefix.Prefix;
 import dev.alexisok.untitledbot.modules.basic.ship.Ship;
+import dev.alexisok.untitledbot.modules.basic.status.Status;
 import dev.alexisok.untitledbot.modules.basic.twenty.TwentyDice;
 import dev.alexisok.untitledbot.modules.rank.Ranks;
 import dev.alexisok.untitledbot.modules.rpg.RPGManager;
-import dev.alexisok.untitledbot.modules.vault.Vault;
 import net.dv8tion.jda.api.EmbedBuilder;
-import net.dv8tion.jda.api.Permission;
-import net.dv8tion.jda.api.entities.Member;
-import net.dv8tion.jda.api.entities.Message;
-import net.dv8tion.jda.api.entities.MessageEmbed;
-import org.jetbrains.annotations.NotNull;
 
 import java.awt.*;
-import java.lang.management.ManagementFactory;
-import java.security.SecureRandom;
-import java.text.SimpleDateFormat;
-import java.util.Arrays;
-import java.util.Objects;
 
 /**
+ * This registers commands or points to all of the commands in the {@link dev.alexisok.untitledbot.modules}
+ * package.
+ * 
  * @author AlexIsOK
  * @since 0.0.1
  */
@@ -37,132 +32,10 @@ public final class CoreCommands {
 	 */
 	public static void registerCoreCommands() {
 		Logger.log("Registering core commands.");
-		
 		//help command
-		CommandRegistrar.register("help", "core.help", ((args, message) -> {
-			EmbedBuilder eb = new EmbedBuilder();
-			EmbedDefaults.setEmbedDefaults(eb, message);
-			try {
-				String returnString = Manual.getHelpPages(args[1]);
-				String embedStr = returnString == null
-						                  ? "Could not find the help page, did you make a typo?"
-						                  : returnString;
-				eb.setColor(returnString == null ? Color.RED : Color.GREEN);
-				eb.addField("Help pages", embedStr, false);
-				return eb.build();
-			} catch(ArrayIndexOutOfBoundsException ignored) {
-				eb.setColor(Color.GREEN);
-				eb.addField("Help pages",
-						"For a list of commands, visit https://github.com/alexisok/untitled-bot/wiki\nFor help with specific " + 
-								"commands, do `help [command]`",
-						false);
-				return eb.build();
-			}
-		}));
-		CommandRegistrar.registerAlias("help", "man", "halp");
-		CommandRegistrar.register("status", "core.stats", (((args, message) -> {
-			EmbedBuilder eb = new EmbedBuilder();
-			EmbedDefaults.setEmbedDefaults(eb, message);
-			String returnString = "```";
-			returnString += "          Version: " + Main.VERSION + "\n";
-			returnString += "       JDA status: " + Main.jda.getStatus() + "\n";
-			returnString += " Available memory: " + Runtime.getRuntime().freeMemory() / 1024 / 1024 + " MB\n";
-			returnString += "     Total memory: " + Runtime.getRuntime().totalMemory() / 1024 / 1024 + " MB\n";
-			returnString += "       Processors: " + Runtime.getRuntime().availableProcessors() + "\n";
-			returnString += "           Uptime: " + ManagementFactory.getRuntimeMXBean().getUptime() + " ms\n";
-			returnString += "  Commands issued: " + CommandRegistrar.getTotalCommands() + "\n";
-			returnString += "   Total messages: " + BotClass.getMessagesSentTotal() + "\n";
-			returnString += "           Guilds: " + Main.jda.getGuilds().size() + "\n";
-			returnString += "            Roles: " + Main.jda.getRoles().size() + "\n";
-			returnString += "            Users: " + Main.jda.getUsers().size() + "\n";
-			returnString += "    Text channels: " + Main.jda.getTextChannels().size() + "\n";
-			returnString += "   Voice channels: " + Main.jda.getVoiceChannels().size() + "\n";
-			returnString += "       Categories: " + Main.jda.getCategories().size() + "\n";
-			returnString += "           Emotes: " + Main.jda.getEmotes().size() + "\n";
-			returnString += "    Shard current: " + message.getJDA().getShardInfo().getShardId() + "\n";
-			returnString += "      Shard total: " + Main.jda.getShardInfo().getShardTotal() + "\n";
-			returnString += "          Plugins: " + CommandRegistrar.registrarSize() + "\n";
-			
-			returnString += "```";
-			
-			eb.setColor(Color.GREEN);
-			eb.addField("Status", returnString, true);
-			return eb.build();
-		})));
-		
-		
-		//the permissions command is very important.  It can be skipped for more security,
-		//but you won't be able to modify command permissions.
-		//Usage: `setperms <user ID|user @|role ID|role @|guild> <permission> <true|false>`
-		CommandRegistrar.register("setperms", "admin", ((args, message) -> {
-			EmbedBuilder eb = new EmbedBuilder();
-			EmbedDefaults.setEmbedDefaults(eb, message);
-			//pre command checks
-			if(message.getAuthor().isBot()) {
-				eb.setColor(Color.RED);
-				eb.addField("Permissions", "Bot users are not allowed to execute this command.", false);
-				return eb.build();
-			}
-			if(!Objects.requireNonNull(message.getMember()).hasPermission(Permission.ADMINISTRATOR)) {
-				eb.setColor(Color.RED);
-				eb.addField("Permissions", "You must be an administrator on the server to execute this command.", false);
-				return eb.build();
-			}
-			try {
-				//make sure that the permission is valid.
-				if(!args[2].matches("^[a-z]([a-z][.]?)+[a-z]$")) {
-					eb.setColor(Color.RED);
-					eb.addField("Permissions", "Please enter a valid command permission.", false);
-					return eb.build();
-				}
-				if(message.getMentionedMembers().size() == 1) {
-					Member mentionedMember = message.getMentionedMembers().get(0);
-					String permission = args[2];
-					boolean allow = args[3].equalsIgnoreCase("true");
-					
-					Vault.storeUserDataLocal(
-							mentionedMember.getId(),
-							message.getGuild().getId(),
-							permission,
-							allow ? "true" : "false"
-					);
-					eb.setColor(Color.GREEN);
-					eb.addField("Permissions", "Permissions updated.", false);
-					return eb.build();
-				} else if(message.getMentionedRoles().size() == 1) {
-					String ID = message.getMentionedRoles().get(0).getId();
-					return getMessageEmbed(args, message, eb, ID);
-				} else if (args[1].matches("^[0-9]+$")) {
-					String memberID = args[1];
-					return getMessageEmbed(args, message, eb, memberID);
-				} else if (args[1].equals("guild")) {
-					String permission = args[2];
-					boolean allow = args[3].equalsIgnoreCase("true");
-					
-					Vault.storeUserDataLocal(
-							null,
-							message.getGuild().getId(),
-							permission,
-							allow ? "true" : "false"
-					);
-					eb.setColor(Color.GREEN);
-					eb.addField("Permissions", "Permissions updated.", false);
-					return eb.build();
-				} else {
-					eb.setColor(Color.RED);
-					eb.addField("Permissions",
-							"Usage: `setperms **<user ID|user @|role ID|role @|guild>** <permission> <true|false>`",
-							false);
-					return eb.build();
-				}
-			} catch(ArrayIndexOutOfBoundsException ignored) {
-				eb.setColor(Color.RED);
-				eb.addField("Permissions",
-						"Usage: `setperms <user ID|user @|role ID|role @|guild> <permission> <true|false>`",
-						false);
-				return eb.build();
-			}
-		}));
+		new Help().onRegister();
+		new Status().onRegister();
+		new Permissions().onRegister();
 		
 		//invite command
 		CommandRegistrar.register("invite", "core.invite", (args, message) -> {
@@ -205,68 +78,19 @@ public final class CoreCommands {
 			eb.addField("title", "works i think", false);
 			return eb.build();
 		});
-		CommandRegistrar.register("set-prefix", "admin", (args, message) -> {
-			EmbedBuilder eb = new EmbedBuilder();
-			EmbedDefaults.setEmbedDefaults(eb, message);
-			
-			if(args.length == 1) {
-				eb.setColor(Color.RED);
-				eb.addField("Prefix", "Usage: set-prefix <prefix>", false);
-				
-				return eb.build();
-			}
-			
-			String prefix = args[1];
-			
-			if(prefix.length() > 3 || prefix.length() < 1) {
-				eb.setColor(Color.RED);
-				eb.addField("Prefix", "Prefix must be one to three characters in length.", false);
-				
-				return eb.build();
-			}
-			
-			Vault.storeUserDataLocal(null, message.getGuild().getId(), "guild.prefix", prefix);
-			
-			eb.setColor(Color.GREEN);
-			eb.addField("Prefix", "Prefix changed to " + prefix + ", however mentioning the bot will work as well.", false);
-			
-			return eb.build();
-		});
-		CommandRegistrar.registerAlias("setperms", "permissions", "perms", "perm", "pr");
+		
+		new Prefix().onRegister();
+		
 		Logger.log("Core commands have been registered.");
 		registerHelp();
 		setDefaults();
 	}
 	
-	/**
-	 * Get the message embed thing that does stuff.
-	 * @param args the arguments for the command
-	 * @param message the message itself
-	 * @param eb the embed builder
-	 * @param ID the ID of the user or role
-	 * @return the new embed thing
-	 */
-	@NotNull
-	private static MessageEmbed getMessageEmbed(String @NotNull [] args, @NotNull Message message, @NotNull EmbedBuilder eb, String ID) {
-		String permission = args[2];
-		boolean allow = args[3].equalsIgnoreCase("true");
-		
-		Vault.storeUserDataLocal(
-				ID,
-				message.getGuild().getId(),
-				permission,
-				allow ? "true" : "false"
-		);
-		eb.setColor(Color.GREEN);
-		eb.addField("Permissions", "Permissions updated.", false);
-		return eb.build();
-	}
 	
 	/**
 	 * Set default permission nodes.
 	 */
 	private static void setDefaults() {
-		CommandRegistrar.setDefaultPermissionForNode("core.help", true);
 		CommandRegistrar.setDefaultPermissionForNode("core.ranks", true);
 		CommandRegistrar.setDefaultPermissionForNode("core.stats", true);
 		CommandRegistrar.setDefaultPermissionForNode("core.invite", true);
@@ -282,18 +106,11 @@ public final class CoreCommands {
 		Manual.setHelpPage("help", "Get help with a specific command.\nUsage: `man <command>`.");
 		Manual.setHelpPage("status", "Get the status of the bot and JVM.");
 		Manual.setHelpPage("shutdown", "Shutdown the bot.\nUsage: `shutdown [code]` where code is the optional exit code.");
-		Manual.setHelpPage("setperms", "Set the permissions of a user, role, or the entire guild.\nUsage: " +
-				                               "setperms <user ID|user @|role ID|role @|guild> <permission> <true|false>");
 		Manual.setHelpPage("invite", "Get the invite link for the bot.");
 		Manual.setHelpPage("about", "much knowledge");
 		Manual.setHelpPage("rank", "Get the current level and XP of a user.\nUsage: " +
 				                           "rank [user @ | user ID]\n" +
 				                           "leave argument blank for your own stats.");
-		Manual.setHelpPage("set-prefix", "Set the prefix for the bot.\n" +
-				                                 "Usage: `set-prefix <prefix>`");
-		CommandRegistrar.registerAliasManual("shutdown", "stop", "exit");
-		CommandRegistrar.registerAliasManual("help", "man", "halp");
-		CommandRegistrar.registerAliasManual("setperms", "permissions", "perms", "perm", "pr");
 	}
 	
 	public static void registerModules() {
