--- src/main/java/dev/alexisok/untitledbot/command/GenericCommandException.java
@@ -1,8 +1,0 @@
-package dev.alexisok.untitledbot.command;
-
-/**
- * @author AlexIsOK
- * @since 0.0.1
- */
-public class GenericCommandException extends RuntimeException {
-}
