+++ src/main/java/dev/alexisok/untitledbot/data/UserDataCouldNotBeObtainedException.java
@@ -4,7 +4,7 @@ package dev.alexisok.untitledbot.data;
  * @author AlexIsOK
  * @since 0.0.1
  */
-public class UserDataCouldNotBeObtainedException extends RuntimeException {
+public final class UserDataCouldNotBeObtainedException extends RuntimeException {
 	public UserDataCouldNotBeObtainedException() {super();}
 	public UserDataCouldNotBeObtainedException(String s) {super(s);}
 }
