+++ src/main/java/dev/alexisok/untitledbot/data/UserDataFileCouldNotBeCreatedException.java
@@ -4,7 +4,7 @@ package dev.alexisok.untitledbot.data;
  * @author AlexIsOK
  * @since 0.0.1
  */
-public class UserDataFileCouldNotBeCreatedException extends RuntimeException {
+public final class UserDataFileCouldNotBeCreatedException extends RuntimeException {
 	public UserDataFileCouldNotBeCreatedException() {super();}
 	public UserDataFileCouldNotBeCreatedException(String s) {super(s);}
 }
