+++ src/main/java/dev/alexisok/untitledbot/modules/basic/ship/Ship.java
@@ -7,7 +7,7 @@ import dev.alexisok.untitledbot.plugin.UBPlugin;
 import net.dv8tion.jda.api.EmbedBuilder;
 import net.dv8tion.jda.api.entities.Message;
 import net.dv8tion.jda.api.entities.MessageEmbed;
-import org.jetbrains.annotations.Nullable;
+import org.jetbrains.annotations.NotNull;
 
 import java.awt.*;
 import java.util.concurrent.ThreadLocalRandom;
@@ -18,10 +18,10 @@ import java.util.concurrent.ThreadLocalRandom;
  * @author AlexIsOK
  * @since 1.0.0
  */
-public class Ship extends UBPlugin {
+public final class Ship extends UBPlugin {
     
     @Override
-    public @Nullable MessageEmbed onCommand(String[] args, Message message) {
+    public @NotNull MessageEmbed onCommand(String[] args, Message message) {
         EmbedBuilder eb = new EmbedBuilder();
         EmbedDefaults.setEmbedDefaults(eb, message);
